+++ dacs7/src/Dacs7Cli/Options/ServerOptions.cs
@@ -5,6 +5,7 @@ namespace Dacs7Cli.Options
     public class ServerOptions : OptionsBase
     {
         public int Port { get; set; }
+        public string BindAddress { get; set; } = Dacs7.Dacs7Server.DefaultBindAddress;
         public List<string> Tags { get; set; }
 
         public string DataProvider { get; set; }